Throw exception if tile_size negative; move print() to utils
