+++ hpx_py11/cpp_code/include/automobile_bits/utils_c.hpp
@@ -1,5 +1,9 @@
 #include <stdio.h>
 #include <tuple>
+#include <stdexcept>
+#include <vector>
+#include <string>
+
 // #include <iostream>
 // #include <hpx/hpx_init.hpp>
 // #include <hpx/future.hpp>
@@ -12,7 +16,10 @@ namespace utils
 
     /// Function to load data from a file into data
     int compute_train_tiles(int n_samples, int n_tile_size);
+
     std::pair<int, int> compute_test_tiles(int m_samples, int n_tiles, int n_tile_size);
+
+    void print(const std::vector<double> &vec, int start = 0, int end = -1, const std::string &separator = " ");
 
     // void print(const std::vector<double> &vec, int start = 0, int end = -1, const std::string &separator = " ");
