+++ hpx_py11/cpp_code/src/gpppy_c.cpp
@@ -46,36 +46,4 @@ namespace gpppy
         return std::move(_data);
     }
 
-    void print(const std::vector<double> &vec, int start, int end, const std::string &separator)
-    {
-        // Convert negative indices to positive
-        if (start < 0)
-            start += vec.size();
-        if (end < 0)
-            end += vec.size() + 1;
-
-        // Ensure the indices are within bounds
-        if (start < 0)
-            start = 0;
-        if (end > vec.size())
-            end = vec.size();
-
-        std::cout << vec.size() << std::endl;
-        // Validate the range
-        if (start >= vec.size() || start >= end)
-        {
-            std::cerr << "Invalid range" << std::endl;
-            return;
-        }
-
-        for (int i = start; i < end; i++)
-        {
-            std::cout << vec[i];
-            if (i < end - 1)
-            {
-                std::cout << separator;
-            }
-        }
-        std::cout << std::endl;
-    }
 }
