+++ hpx_py11/cpp_code/src/utils_c.cpp
@@ -1,5 +1,7 @@
 #include "../include/automobile_bits/utils_c.hpp"
 
+#include <cstdio>
+#include <iostream>
 // #include <iostream>
 // #include <hpx/hpx.hpp>
 // #include <hpx/future.hpp>
@@ -12,10 +14,17 @@ namespace utils
 {
     int compute_train_tiles(int n_samples, int n_tile_size)
     {
-        std::size_t _n_samples = static_cast<std::size_t>(n_samples);
-        std::size_t _n_tile_size = static_cast<std::size_t>(n_tile_size);
-        std::size_t _n_tiles = _n_samples / _n_tile_size;
-        return static_cast<int>(_n_tiles);
+        if (n_tile_size > 0)
+        {
+            std::size_t _n_samples = static_cast<std::size_t>(n_samples);
+            std::size_t _n_tile_size = static_cast<std::size_t>(n_tile_size);
+            std::size_t _n_tiles = _n_samples / _n_tile_size;
+            return static_cast<int>(_n_tiles);
+        }
+        else
+        {
+            throw std::runtime_error("Error: Please specify a valid value for train_tile_size.\n");
+        }
     }
 
     std::pair<int, int> compute_test_tiles(int n_test, int n_tiles, int n_tile_size)
@@ -37,5 +46,38 @@ namespace utils
             m_tile_size = _n_tile_size;
         }
         return {static_cast<int>(m_tiles), static_cast<int>(m_tile_size)};
+    }
+
+    void print(const std::vector<double> &vec, int start, int end, const std::string &separator)
+    {
+        // Convert negative indices to positive
+        if (start < 0)
+            start += vec.size();
+        if (end < 0)
+            end += vec.size() + 1;
+
+        // Ensure the indices are within bounds
+        if (start < 0)
+            start = 0;
+        if (end > vec.size())
+            end = vec.size();
+
+        std::cout << vec.size() << std::endl;
+        // Validate the range
+        if (start >= vec.size() || start >= end)
+        {
+            std::cerr << "Invalid range" << std::endl;
+            return;
+        }
+
+        for (int i = start; i < end; i++)
+        {
+            std::cout << vec[i];
+            if (i < end - 1)
+            {
+                std::cout << separator;
+            }
+        }
+        std::cout << std::endl;
     }
 }
